add bytecode compilation & vm interpretation of add, neg, const, output instructions

+++ src/vm/value.rs
@@ -1,1 +1,1 @@
-pub type Value = f32;
+pub type Value = i32;
